Add files via upload
I updated the groups code and made it activate the door motor with the face detection. I added a delay to the opening so that it stays open for longer and tested it before I left for break. I also picked up Matt's 3d print of the door while he was gone. Our project is almost complete! I just need to finish up the box and attaching the door. I fear I might have trouble attaching the door because this one is heavier than the paper and I'm scared it might fall apart. I need to find a way to make supports for it somehow.